Wrap tsnet Listener with tsnetServerListener
Everytime we get a listener we increment the refernce on a caddy
UsagePool. With this change, wrap the listener in a custom type that
will decrement the UsagePool when we close the listener. When all
listeners are closed we will call Destruct on the tsnet Server which
will Close the server and should cleanup potential references in the
tailscale control plane

Signed-off-by: Connor Kelly <connor.r.kelly@gmail.com>