set correct tailnet name in authentication provider
The MagicDNS suffix is what we now refer to as the "Tailnet Name":
https://tailscale.com/kb/1217/tailnet-name. This code was previously
using what we now call the org name, which used to be embedded in the
MagicDNS name before we used ts.net cert domains.

Also add tailnet name to the output in the proxyauth example config.

Fixes #29

Signed-off-by: Will Norris <will@tailscale.com>