Dockerfile: update git
Apparently the old version isn't even available in the package server
anymore?  That's pretty annoying, and kinda makes me question pinning
the version.  For now, I'll just bump it, but might just need to drop it
entirely and take the latest version.  We're not using it for much
anyway.

Also fix a few Dockerfile linting issues.

Updates #cleanup

Change-Id: Ia101a4a3005adb9118051b3416f5a64a4a45987d
Signed-off-by: Will Norris <will@tailscale.com>